chore: rename inkbunny.description.ts to inkbunny-description-converter.ts
